moving check_provider_availability task to scheduled_tasks.py, fixing reference to DataProviderStatusInline
